+++ src/main/java/org/styly/acanus/item/ImbuableDarkChestplateArmorItem.java
@@ -1,0 +1,55 @@
+package org.styly.acanus.item;
+
+import io.redspace.ironsspellbooks.api.spells.IPresetSpellContainer;
+import io.redspace.ironsspellbooks.api.spells.ISpellContainer;
+import io.redspace.ironsspellbooks.item.armor.ExtendedArmorItem;
+import io.redspace.ironsspellbooks.item.armor.IronsExtendedArmorMaterial;
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ArmorItem;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
+import org.styly.acanus.registry.ModEffects;
+
+public abstract class ImbuableDarkChestplateArmorItem extends ExtendedArmorItem implements IPresetSpellContainer {
+    public ImbuableDarkChestplateArmorItem(IronsExtendedArmorMaterial material, ArmorItem.Type type, Item.Properties settings) {
+        super(material, type, settings);
+    }
+
+    private boolean hasFullSuitOfArmorOn(Player player) {
+        ItemStack boots = player.getInventory().getArmor(0);
+        ItemStack leggings = player.getInventory().getArmor(1);
+        ItemStack breastplate = player.getInventory().getArmor(2);
+        ItemStack helmet = player.getInventory().getArmor(3);
+
+        return !helmet.isEmpty() && !breastplate.isEmpty()
+                && !leggings.isEmpty() && !boots.isEmpty();
+    }
+
+    @Override
+    public void onArmorTick(ItemStack stack, Level world, Player player) {
+        if (!world.isClientSide()) {
+            if (hasFullSuitOfArmorOn(player)) {
+                if(world.isNight()){
+                    player.addEffect(new MobEffectInstance(ModEffects.FULL_MOON.get(), 60));
+                }
+            }
+        }
+    }
+
+    @Override
+    public void initializeSpellContainer(ItemStack itemStack) {
+        if (itemStack == null) {
+            return;
+        }
+
+        if (itemStack.getItem() instanceof ArmorItem armorItem && armorItem.getType() == ArmorItem.Type.CHESTPLATE) {
+            if (!ISpellContainer.isSpellContainer(itemStack)) {
+                var spellContainer = ISpellContainer.create(1, true, true);
+                spellContainer.save(itemStack);
+            }
+        }
+    }
+}
+