+++ src/main/java/org/styly/acanus/spells/ManaCombustion.java
@@ -11,6 +11,7 @@ import io.redspace.ironsspellbooks.api.util.CameraShakeManager;
 import io.redspace.ironsspellbooks.api.util.Utils;
 import io.redspace.ironsspellbooks.capabilities.magic.MagicManager;
 import io.redspace.ironsspellbooks.damage.DamageSources;
+import io.redspace.ironsspellbooks.entity.spells.fireball.MagicFireball;
 import io.redspace.ironsspellbooks.particle.ShockwaveParticleOptions;
 import io.redspace.ironsspellbooks.registries.SoundRegistry;
 import net.minecraft.core.particles.ParticleTypes;
@@ -34,8 +35,8 @@ import java.util.List;
 import java.util.Optional;
 
 @AutoSpellConfig
-public class ManaSmite extends AbstractSpell {
-    private final ResourceLocation spellId = new ResourceLocation(Arcanus.MODID, "mana_smite");
+public class ManaCombustion extends AbstractSpell {
+    private final ResourceLocation spellId = new ResourceLocation(Arcanus.MODID, "mana_combustion");
 
     @Override
     public List<MutableComponent> getUniqueInfo(int spellLevel, LivingEntity caster) {
@@ -49,7 +50,7 @@ public class ManaSmite extends AbstractSpell {
             .setCooldownSeconds(15)
             .build();
 
-    public ManaSmite() {
+    public ManaCombustion() {
         this.manaCostPerLevel = 15;
         this.baseSpellPower = 25;
         this.spellPowerPerLevel = 3;
@@ -70,7 +71,7 @@ public class ManaSmite extends AbstractSpell {
 
     @Override
     public CastType getCastType() {
-        return CastType.INSTANT;
+        return CastType.LONG;
     }
 
     @Override
@@ -95,24 +96,19 @@ public class ManaSmite extends AbstractSpell {
 
     @Override
     public void onCast(Level level, int spellLevel, LivingEntity entity, CastSource castSource, MagicData playerMagicData) {
-        float radius = 5f;
+        float radius = 15f;
         Vec3 smiteLocation = Utils.moveToRelativeGroundLevel(level, entity.getEyePosition().add(entity.getForward().multiply(1.35f, 0, 1.35f)), 1);
         MagicManager.spawnParticles(level, new ShockwaveParticleOptions(SchoolRegistry.ELDRITCH.get().getTargetingColor(), radius * 2, true), smiteLocation.x, smiteLocation.y, smiteLocation.z, 1, 0, 0, 0, 0, true);
         MagicManager.spawnParticles(level, ParticleTypes.ELECTRIC_SPARK, smiteLocation.x, smiteLocation.y, smiteLocation.z, 50, 0, 0, 0, 1, false);
         CameraShakeManager.addCameraShake(new CameraShakeData(42, smiteLocation, 10));
-        var entities = level.getEntities(entity, AABB.ofSize(smiteLocation, radius * 2, radius * 4, radius * 2));
-        for (Entity targetEntity : entities) {
-            //double distance = targetEntity.distanceToSqr(smiteLocation);
-            if (targetEntity.isAlive() && targetEntity.isPickable() && Utils.hasLineOfSight(level, smiteLocation.add(0, 1, 0), targetEntity.getBoundingBox().getCenter(), true)) {
-                if (DamageSources.applyDamage(targetEntity, getDamage(spellLevel, entity), this.getDamageSource(entity))) {
-                    int i = EnchantmentHelper.getFireAspect(entity);
-                    if (i > 0) {
-                        targetEntity.setSecondsOnFire(i * 4);
-                    }
-                    EnchantmentHelper.doPostDamageEffects(entity, targetEntity);
-                }
-            }
-        }
+        Vec3 origin = entity.getEyePosition();
+        MagicFireball fireball = new MagicFireball(level, entity);
+
+        fireball.setDamage(getDamage(spellLevel, entity));
+        fireball.setExplosionRadius(15);
+        fireball.setPos(origin.add(entity.getForward()).subtract(0, fireball.getBbHeight() / 2, 0));
+        fireball.shoot(new Vec3(0,-1,0));
+        level.addFreshEntity(fireball);
         super.onCast(level, spellLevel, entity, castSource, playerMagicData);
     }